requirements.txt
fastapi==0.136.0
huggingface_hub==0.34.4
numpy<2.0
opencv-python-headless==4.10.0.84
pillow==11.2.1
python-multipart==0.0.20
ultralytics==8.3.177
uvicorn==0.34.3
